converted other constants
