+++ src/js/components/overview/charts/Chart.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { chartTypes } from '../../../constants/overviewConstants';
 import BentoBarChart from './BentoBarChart';
 import BentoPie from './BentoPie';
 
@@ -7,11 +8,11 @@ const CHART_HEIGHT = 300;
 const Chart = ({ chartType, data, title, units }) => {
   const renderChartSwitch = () => {
     switch (chartType) {
-      case 'bar':
+      case chartTypes.BAR:
         return (
           <BentoBarChart data={data} height={CHART_HEIGHT} units={units} />
         );
-      case 'pie':
+      case chartTypes.PIE:
         return <BentoPie data={data} height={CHART_HEIGHT} />;
       default:
         return <p>chart type doesnt exists</p>;
