+++ src/js/constants/overviewConstants.js
@@ -1,13 +1,7 @@
-const { CLIENT, HOST, DEBUG, MAX_QUERY_PARAMETERS } = process.env;
-
-export const debug = DEBUG.toLowerCase() === 'true';
-export const client = CLIENT;
-export const maxQueryParameters = MAX_QUERY_PARAMETERS;
-
-export const configUrl = HOST + '/config';
-export const publicOverviewUrl = HOST + '/overview';
-export const queryableFieldsUrl = HOST + '/fields';
-export const katsuUrl = HOST + '/katsu';
+export const chartTypes = {
+  BAR: 'bar',
+  PIE: 'pie',
+};
 
 // Bento-web colours
 export const COLORS = [
